+++ apps/web/src/app/server/layout.tsx
@@ -1,5 +1,9 @@
 import { ThemeProvider } from "next-server-theme/dev/server.ts";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
-    return <ThemeProvider>{children}</ThemeProvider>;
+    return (
+        <ThemeProvider themes={["dark", "light"]} defaultTheme="bhj">
+            {children}
+        </ThemeProvider>
+    );
 }