+++ packages/next-server-theme/src/react/util/setBackgroundTheme.ts
@@ -3,9 +3,18 @@ import type { Theme } from "../types";
 export function setBackgroundTheme(
     theme: Theme,
     element: string = "html",
-    attributes: string | string[] = "class"
+    attributes: string | string[] = "class",
+    themes?: Theme[]
 ) {
+    console.log(themes);
     [attributes].flat().forEach(attribute => {
-        document.querySelector(element)?.setAttribute(attribute, theme);
+        if (attribute === "class" && themes) {
+            console.log("class with themes");
+            document.querySelector(element)?.classList.remove(...themes);
+            document.querySelector(element)?.classList.add(theme);
+        } else {
+            console.log("nah");
+            document.querySelector(element)?.setAttribute(attribute, theme);
+        }
     });
 }